feat: add app start and browser opening
