+++ package.json
@@ -3,9 +3,8 @@
   "version": "1.0.0",
   "description": "Template for the WebSocket remote control task",
   "main": "index.js",
-  "type": "module",
   "scripts": {
-    "start": "node ./index.js",
+    "start": "tsc && node dist/index.js",
     "start:dev": "nodemon ./index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
