+++ src/http_server/index.ts
@@ -1,0 +1,38 @@
+import * as fs from 'fs';
+import * as path from 'path';
+import * as http from 'http';
+import { exec } from 'child_process';
+
+export const httpServer = http.createServer(function (req, res) {
+  const __dirname = path.resolve(path.dirname(''));
+  const file_path =
+    __dirname + (req.url === '/' ? '/front/index.html' : '/front' + req.url);
+  fs.readFile(file_path, function (err, data) {
+    if (err) {
+      res.writeHead(404);
+      res.end(JSON.stringify(err));
+      return;
+    }
+    res.writeHead(200);
+    res.end(data);
+  });
+});
+
+export function openBrowser(url: string) {
+  const platform = process.platform;
+  let command;
+
+  if (platform === 'win32') {
+    command = `start ${url}`;
+  } else if (platform === 'darwin') {
+    command = `open ${url}`;
+  } else {
+    command = `xdg-open ${url}`;
+  }
+
+  exec(command, (err) => {
+    if (err) {
+      console.error(`Failed to open browser: ${err}`);
+    }
+  });
+}
